# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateCaseRequest": {
        "additionalProperties": false,
        "description": "A de-identified teacher case without learner identity fields.",
        "properties": {
          "deidentified_attestation": {
            "default": false,
            "title": "Deidentified Attestation",
            "type": "boolean"
          },
          "observed_work": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Observed Work",
            "type": "string"
          },
          "problem": {
            "$ref": "#/components/schemas/SignedProblemDTO"
          },
          "provenance_record_id": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance Record Id"
          },
          "source_tier": {
            "enum": [
              "authentic",
              "synthetic",
              "mixed",
              "published_exemplar",
              "educator_authored",
              "custom"
            ],
            "title": "Source Tier",
            "type": "string"
          }
        },
        "required": [
          "source_tier",
          "problem",
          "observed_work"
        ],
        "title": "CreateCaseRequest",
        "type": "object"
      },
      "CreateCaseResponse": {
        "additionalProperties": false,
        "description": "Opaque identifiers for the newly owned case and workflow.",
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "workflow_id": {
            "format": "uuid",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "workflow_id"
        ],
        "title": "CreateCaseResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "COGNISECT API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases": {
      "post": {
        "operationId": "create_case_route_v1_cases_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "pattern": "^[\\x21-\\x7e]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cognisect_owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cognisect Owner"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerBootstrapResponse"
                }
              }
            },
            "description": "Owner session initialized before educational mutation"
          },
          "429": {
            "description": "Fixed-window request quota exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the current quota window expires",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Create Case Route"
      }
    },
    "/v1/cases/{case_id}/analysis": {
      "post": {
        "operationId": "analyze_case_route_v1_cases__case_id__analysis_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "pattern": "^[\\x21-\\x7e]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cognisect_owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cognisect Owner"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Fixed-window request quota exceeded",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the current quota window expires",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Analyze Case Route"
      }
    },
    "/v1/respond/{token}": {
      "get": {
        "operationId": "get_learner_probe_route_v1_respond__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerProbeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learner Probe Route"
      },
      "post": {
        "operationId": "submit_learner_route_v1_respond__token__post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "pattern": "^[\\x21-\\x7e]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerReceiptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Learner Route"
      }
    },
    "/v1/workflows/{workflow_id}": {
      "delete": {
        "operationId": "delete_workflow_route_v1_workflows__workflow_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "pattern": "^[\\x21-\\x7e]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cognisect_owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cognisect Owner"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workflow Route"
      },
      "get": {
        "operationId": "get_workflow_route_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cognisect_owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cognisect Owner"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow Route"
      }
    },
...
  }
}